Implement improved language summary breakdown
Previously this functionality relied on the user having
`github-linguist` installed (which, for me, I would install and it
would always break due to weird versions of Ruby).  It also required
me to have the language colours YAML file stored away in a specific
directory.

With the discovery of hyperpolyglot, this has been greatly improved!