Update dependencies (including clap)
No functionality changes, but some back-end changes, especially to
clap, which I updated from v2 to v4.  v4 introduced a
whole bunch of breaking changes, but I have been working

While I have fixed all compiler errors to do with the breaking
changes, I may need to go through and do some testing to ensure that
all of the app's functionality is not panicing.

I really like clap's new API.  It seems really logical.  In v2, you
can tell that a lot of the functionality that was present was made on
top of an already existing API, and nobody wanted to change anything
fundamentally with clap.  The new verson does away with methods such
as `min_args`, `max_args`, `multiple`, and `takes_value`, with
essentially one replacement: `num_args`, which takes a number or
range.  For example, if I make an argument with `num_args(0)`, this
_implies_ `multiple(false)`, `min_args(0)`, and `max_args(0)`.  If I
make an argument with `num_args(1..=2)`, this implies
`multiple(true)`, `min_args(1)`, and `max_args(2)`.  If I make an
argument with `num_args(0..=1)`, this implies `multiple(false)`,
`min_args(0)`, and `max_args(1)`.  The new version seems to condense a
whole lot of other logic, too, into a really clean API.