Decouple git commit collection for log
We essentially just redirect the output of `git log` to stdout when we
run `gl` without arguments (where it prints its logs).  However, with
a little more parsing, we can decouple the commit collection step so
that we can use these in other parts of the codebase.

It's still rather messy (the whole codebase could probably do with a
rewrite---see #2 and #8---but this is slightly better, for now.