Make commit bound optional in git_log
As git_log was created for logging, it has builtin functionality to
limit the number of logs you want to display.  However, we need all
logs for contributions.  So I

I should also allow all logs to be printed anyway, because sometimes I
want that feature so I just put a really big number in.  I will add
this to the issue tracker: #21