Refine contributions.rs with more logical structs and use of methods
Since moving commit functionality to a separate submodule, we can use
this in contributions.rs, in preparation for the commit graph.  It's
still pretty messy but there were a few ways we could refine the logic
in contributions.rs; mostly using better struct/method names and
rewriting parts of the methods as a result.

We also, of course, set up some date-/commit-based software
infrastructure so that we can work with the contributions graph soon.