Use less common delimiter for commit meta info!  v2.3.1
When we obtain commits, we are required to extract some meta
information from them.  Previously the format was:

    <commit to display> {{<meta information>}}

However, this *panics* when double curly braces are used in the commit
message.  I figured this wouldn't be common, but the week after
implementing v2.3.0 I found a situation in which this occurred.

As a result, we need a better delimiter.  The new format is as
follows:

    <commit to display>⸺<meta information>

(The delimiter is now a two em-dash.)  This is far less likely to be
used in a commit message (at least in the head of one), and is a
semantically meaningful delimiter as well.

Hopefully we don't run into any further issues.