Only allow one functional option at a time!  v3.0.0
The primary change here is using an argument group to collect
different arguments (which is then flattened in the primary Cli struct
that derives the argument parser).  Importantly, this argument group
contains the options that do functional things.  It also specifies
`multiple = false`, meaning that we can only run `gl` with one
functional option at a time.  (Note that there are still non-
functional options that you can use, such as `--rev` and `--all` which
will change the way `gl`'s base functionality (pretty logging) will
print.)  This change closes #28.

I have considered this a breaking change because you were previously
able to run the program with multiple options specified, but this is
no longer allowed.

While we are making changes that can be considered breaking changes,
we take the opportunity to do a few more outrageous things:
  - Rename `--commit-count-when` to `--commit-count-at`;
  - Deprecate `--global` (as it's not been working for a long time and
    I don't think I will ever need the functionality);
  - Run clippy on the newest version of the Rust toolchain and perform
    any suggested changes.  (I really need to get around to adding
    CI/CD; #30.)

The whole change goes a good way towards addressing #2.