Use unique quotes in git log metadata collection !  v3.0.2
Previously we had metadata formatted at the end of the log in the
following manner:

    {delim}<field>: '{value}'

However, this caused issues because for the value regex we would
simple look for any character except the closing delimiter '.  If an
email had an apostrophe in in (which is allowed---thanks @alok-dh),
this would break.  Instead, we can use unique quote characters (whose
closing character is distinct from the opening character) to work
around this issue:

    {delim}<field>: “{value}”

In doing this, I moved some of the patterns to be statically compiled
with lazy_static.