Use more unique quotes in git log metadata collection!  v3.1.2
Previously we had metadata formatted at the end of the log in the
following manner:

    {delim}<field>: '{value}'

Then, when this failed with users' emails containing apostrophes,  we
tried:

    {delim}<field>: “{value}”

Which worked for a while.

However, this causes issues because a user who committed to Mozilla's
monolith repository had username `“mihai.carare.dev@gmail.com”`.

We need to use even *more* unique quotes---unique-er quotes---to
delineate what part of the log contains the values we need to extract
for pretty printing.  So, we work around with some fancy looking box
drawing characters:

    {delim}<field>: ╠{value}╣

If anyone ever uses these in their email, I'm going to eat a gun.

See also v3.0.2 [1] / 965093e.

Closes #41

[1]: https://github.com/jakewilliami/gl/releases/tag/v3.0.2